+++ packages/ui-library/src/components/Chips/Chips.tsx
@@ -8,12 +8,12 @@ import type { TClickEventType } from '../../types/globalTypes';
 import { ChipCustomType } from './types';
 import { COLOR_MAPPING, LEFT_ICON_SIZE_MAPPING, RIGHT_ICON_SIZE_MAPPING, TEXT_SIZE_MAPPING } from './consts';
 import { Text } from '../Text';
-import IconDismissCircleFilled from '../SVGIcons/IconDismissCircleFilled';
+import IconDismissCircle from '../SVGIcons/IconDismissCircle';
 
 export const Chips = (props: TChipsProps): ReactElement => {
   const {
     type = 'filled',
-    color = 'primary',
+    color = 'default',
     size = 'large',
     text,
     disabled,
@@ -25,7 +25,11 @@ export const Chips = (props: TChipsProps): ReactElement => {
     dataId = '',
   } = props;
 
-  const customType = disabled ? 'disabled' : type == ChipCustomType.filled ? ChipCustomType.inverse : color;
+  const customType = disabled
+    ? 'disabled'
+    : type == ChipCustomType.filled
+      ? ChipCustomType.inverse
+      : COLOR_MAPPING[color];
 
   const handleClick = (event: TClickEventType) => {
     event.stopPropagation();
@@ -41,15 +45,15 @@ export const Chips = (props: TChipsProps): ReactElement => {
         <leftIconProps.Component
           dataId={`${dataId}-icon`}
           size={RIGHT_ICON_SIZE_MAPPING[size]}
-          type={type == ChipCustomType.filled ? 'inverse' : COLOR_MAPPING[customType]}
-          className={'chips__icon'}
+          type={customType}
+          className="chips__icon"
           {...leftIconProps}
         />
       ) : null}
       {text ? (
         <Text
           dataId={`${dataId}-text`}
-          type={type == ChipCustomType.filled ? 'inverse' : COLOR_MAPPING[customType]}
+          type={customType}
           size={TEXT_SIZE_MAPPING[size]}
           className="chips__label text-truncate"
         >
@@ -57,10 +61,10 @@ export const Chips = (props: TChipsProps): ReactElement => {
         </Text>
       ) : null}
       {withAction && (
-        <IconDismissCircleFilled
+        <IconDismissCircle
           dataId={`${dataId}-icon`}
           size={LEFT_ICON_SIZE_MAPPING[size]}
-          type={type == ChipCustomType.filled ? 'inverse' : COLOR_MAPPING[customType]}
+          type={customType}
           className="chips__delete ml-4"
           onClick={handleClick}
         />